+++ Verizon/src/main/java/com/mopub/mobileads/VerizonBanner.java
@@ -285,15 +285,17 @@ public class VerizonBanner extends CustomEventBanner {
         public void onLoaded(final InlineAdFactory inlineAdFactory, final InlineAdView inlineAdView) {
             MoPubLog.log(LOAD_SUCCESS, ADAPTER_NAME);
 
-            final CreativeInfo creativeInfo = verizonInlineAd == null ? null : verizonInlineAd.getCreativeInfo();
-            MoPubLog.log(CUSTOM, ADAPTER_NAME, "Verizon creative info: " + creativeInfo);
+            verizonInlineAd = inlineAdView;
 
             VerizonUtils.postOnUiThread(new Runnable() {
 
                 @Override
                 public void run() {
+                    final CreativeInfo creativeInfo = verizonInlineAd == null ? null : verizonInlineAd.getCreativeInfo();
+                    MoPubLog.log(CUSTOM, ADAPTER_NAME, "Verizon creative info: " + creativeInfo);
+
                     if (internalView != null) {
-                        internalView.addView(inlineAdView);
+                        internalView.addView(verizonInlineAd);
                     }
 
                     if (listener != null) {