+++ Verizon/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 1.1.3.1
+    * Add support for Verizon native.
+    * Add support for Verizon rewarded video.
+
   * 1.1.3.0
     * This version of the adapters has been certified with Verizon 1.1.3.